Improved admin dashboard performance for large numbers of users
With around 6500 users in LDAP, the user and project admin panels become
unusable.

Modified the project members workflow step layout so that it uses a
simple select control for adding new members instead of a group of DOM
elements for every single user.  This has cut down the page load time
to a 1/3 of what it was before and is now far more responsive.

For the users panel, we simply exclude LDAP users from the list since
they cannot be modified (unlike users defined in keystone) so there is
little value including them there.