Address feedback from Chris.
Don't bail completely if an instance is missing OS-EXT-SRV-ATTR:host,
but instead just ignore it, show an error, and try to show the remaining
instances.

When details for a flavor or project need to be fetched a second time,
log a warning instead of showing a pop-up warning message, to be less
annoying...

Include ephemeral disk usage in calculations.

Remove some redundant code and also clean up a little bit.